added dynamic relations by function mapping linked fields and restructured the code to split resolvers and shcemas into separated files
